requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.4
fsspec==2025.5.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
PyYAML==6.0.2
requests==2.32.4
rubicon-objc==0.5.1
scipy==1.13.1
six==1.17.0
soupsieve==2.7
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.157
ultralytics-thop==2.0.14
urllib3==2.5.0
Werkzeug==3.1.3
wikipedia==1.4.0
zipp==3.23.0
